oxish: move DefaultStore and SystemStore into unix
