+++ oxish/src/unix/mod.rs
@@ -1,21 +1,25 @@
-use core::{cmp::Ordering, mem::MaybeUninit};
+use core::{cmp::Ordering, ffi::c_char, mem::MaybeUninit, ops::ControlFlow};
 #[cfg(coverage)]
 use std::env;
 use std::{
-    ffi::CString,
-    io::{self, IoSlice, IoSliceMut},
+    ffi::{CStr, CString, OsStr},
+    fs::File,
+    io::{self, IoSlice, IoSliceMut, Read},
     os::{
         fd::{AsFd, OwnedFd},
-        unix::{ffi::OsStrExt, net::UnixStream},
+        unix::{ffi::OsStrExt, fs::MetadataExt, net::UnixStream},
     },
+    path::PathBuf,
     process::Stdio,
 };
 
+use libc::{_SC_GETPW_R_SIZE_MAX, ERANGE, getpwnam_r, getpwuid_r, sysconf};
 use proto::{
-    Decoded, Encode, ReadState, ServerHostKey, SessionHostKey, WriteState, crypto::CryptoProvider,
-    key_exchange::Rekey,
+    Decoded, Encode, ReadState, ServerHostKey, SessionHostKey, WriteState, auth::AuthorizedKey,
+    crypto::CryptoProvider, key_exchange::Rekey,
 };
 use rustix::{
+    fs::{Mode, OFlags, openat},
     io::FdFlags,
     net::{
         RecvAncillaryBuffer, RecvAncillaryMessage, RecvFlags, SendAncillaryBuffer,
@@ -26,10 +30,16 @@ use tokio::{
     net::TcpStream,
     process::{Child, Command},
 };
-use tracing::debug;
+use tracing::{debug, error, warn};
 use zeroize::Zeroizing;
 
-use crate::{Connection, Error, Server, Session, SessionState, User, session::Channels};
+use crate::{
+    Connection, Error, SessionState,
+    authentication::{CachedUser, SingleUser, User, UserStore, Username},
+    server::Server,
+    session::Channels,
+    session::Session,
+};
 
 mod terminal;
 pub(crate) use terminal::Terminal;
@@ -311,6 +321,290 @@ pub fn resume(provider: &'static dyn CryptoProvider) -> Result<Session<TcpStream
         channels: Channels::default(),
         post_quantum_kx,
     })
+}
+
+/// Default [`UserStore`] implementation
+///
+/// Uses the system database when running as root, and a single-user store otherwise.
+pub struct DefaultStore(());
+
+impl DefaultStore {
+    /// Construct a new [`DefaultStore`] from the current process's effective UID
+    #[expect(clippy::new_ret_no_self)]
+    pub fn new(provider: &dyn CryptoProvider) -> Result<Box<dyn UserStore>, Error> {
+        // SAFETY: `geteuid()` takes no arguments, cannot fail and has no preconditions.
+        Ok(match unsafe { libc::geteuid() } {
+            0 => {
+                debug!("using system user store");
+                Box::new(SystemStore) as Box<dyn UserStore>
+            }
+            uid => {
+                debug!(uid, "using single-user store");
+                let data = UserLookup::Id(uid).resolve()?;
+                let keys = SystemStore.keys(&data, provider);
+                Box::new(SingleUser(CachedUser { data, keys }))
+            }
+        })
+    }
+}
+
+/// User store backed by the system database
+pub(crate) struct SystemStore;
+
+impl UserStore for SystemStore {
+    fn lookup(&self, name: Username) -> Option<User> {
+        match UserLookup::Name(name).resolve() {
+            Ok(user) => Some(user),
+            Err(error) => {
+                error!(%error, "failed to get user information");
+                None
+            }
+        }
+    }
+
+    fn keys(&self, user: &User, provider: &dyn CryptoProvider) -> Vec<AuthorizedKey> {
+        let home_dir = &user.home_dir;
+        let home = match File::open(home_dir) {
+            Ok(file) => file,
+            Err(error) => {
+                warn!(%error, ?home_dir, "failed to open home directory");
+                return Vec::new();
+            }
+        };
+
+        match check_permissions(&home, user.id, "home directory") {
+            ControlFlow::Continue(()) => {}
+            ControlFlow::Break(()) => {
+                warn!(?home_dir, "bad permissions on home directory");
+                return Vec::new();
+            }
+        };
+
+        let result = openat(
+            &home,
+            ".ssh",
+            OFlags::RDONLY | OFlags::DIRECTORY | OFlags::CLOEXEC,
+            Mode::empty(),
+        );
+        let ssh_dir = match result {
+            Ok(fd) => File::from(fd),
+            Err(error) => {
+                warn!(%error, ?home_dir, "failed to open .ssh directory");
+                return Vec::new();
+            }
+        };
+
+        match check_permissions(&ssh_dir, user.id, ".ssh directory") {
+            ControlFlow::Continue(()) => {}
+            ControlFlow::Break(()) => {
+                warn!(?home_dir, "bad permissions on .ssh directory");
+                return Vec::new();
+            }
+        };
+
+        let result = openat(
+            &ssh_dir,
+            "authorized_keys",
+            OFlags::RDONLY | OFlags::CLOEXEC,
+            Mode::empty(),
+        );
+        let mut key_file = match result {
+            Ok(fd) => File::from(fd),
+            Err(error) => {
+                warn!(%error, ?home_dir, "failed to open authorized keys file");
+                return Vec::new();
+            }
+        };
+
+        match check_permissions(&key_file, user.id, "authorized keys file") {
+            ControlFlow::Continue(()) => {}
+            ControlFlow::Break(()) => {
+                warn!(?home_dir, "bad permissions on authorized keys file");
+                return Vec::new();
+            }
+        };
+
+        let mut contents = String::new();
+        if let Err(error) = key_file.read_to_string(&mut contents) {
+            warn!(%error, ?home_dir, "failed to read authorized keys file");
+            return Vec::new();
+        };
+
+        let mut keys = Vec::new();
+        for (line, key) in contents.lines().enumerate() {
+            match AuthorizedKey::from_str(key, provider) {
+                Some(key) => keys.push(key),
+                None => debug!(line = line + 1, "no valid authorized key found on line"),
+            }
+        }
+
+        keys
+    }
+
+    fn drop_privileges(&self) -> bool {
+        true
+    }
+}
+
+#[derive(Debug)]
+enum UserLookup {
+    Name(Username),
+    Id(u32),
+}
+
+impl UserLookup {
+    fn resolve(self) -> Result<User, Error> {
+        /// Upper bound on the buffer used to hold the passwd entry
+        const MAX_BUF_LEN: usize = 1_048_576;
+
+        // SAFETY: `sysconf()` only reads its integer argument and has no other preconditions.
+        let buf_len = match unsafe { sysconf(_SC_GETPW_R_SIZE_MAX) } {
+            -1 => 1024,
+            n => (n as usize).clamp(1024, MAX_BUF_LEN),
+        };
+
+        let c_name = match &self {
+            Self::Name(name) => Some(CString::new(&**name).map_err(|_| Error::InvalidUsername)?),
+            Self::Id(_) => None,
+        };
+
+        let mut buf = vec![0u8; buf_len];
+        // SAFETY: `passwd` is a plain C struct of integers and pointers, for which
+        // all-zeros (including null pointers) is a valid bit pattern.
+        let mut pwd = unsafe { core::mem::zeroed() };
+        let mut result = core::ptr::null_mut();
+
+        // A passwd entry can exceed the initial buffer size (a long GECOS field is
+        // enough); `ERANGE` means the buffer was too small, so grow it and try again,
+        // up to a cap (like the `getgrouplist()` loop in `server.rs`).
+        let ret = loop {
+            let ret = match (&self, &c_name) {
+                (Self::Name(_), Some(c_name)) => unsafe {
+                    // SAFETY: `c_name` is a valid null-terminated C string, `pwd` and `result` are
+                    // valid for writes, and the buffer pointer and length describe the live
+                    // allocation in `buf`.
+                    getpwnam_r(
+                        c_name.as_ptr(),
+                        &mut pwd,
+                        buf.as_mut_ptr().cast::<c_char>(),
+                        buf.len(),
+                        &mut result,
+                    )
+                },
+                (Self::Id(id), _) => unsafe {
+                    // SAFETY: `pwd` and `result` are valid for writes, and the buffer pointer
+                    // and length describe the live allocation in `buf`.
+                    getpwuid_r(
+                        *id,
+                        &mut pwd,
+                        buf.as_mut_ptr().cast::<c_char>(),
+                        buf.len(),
+                        &mut result,
+                    )
+                },
+                (Self::Name(_), None) => {
+                    unreachable!("`c_name` is set for lookups by name")
+                }
+            };
+
+            if ret != ERANGE || buf.len() >= MAX_BUF_LEN {
+                break ret;
+            }
+
+            buf.resize(Ord::min(buf.len() * 2, MAX_BUF_LEN), 0);
+        };
+
+        let name = match self {
+            Self::Name(name) => name,
+            Self::Id(_) => match (ret, result.is_null(), pwd.pw_name.is_null()) {
+                // SAFETY: `ret` is 0 and `result` is non-null, so the `pwd.pw_name` points to a
+                // null-terminated C string stored in `buf`, which is still alive.
+                (0, false, false) => Username::try_from(unsafe { CStr::from_ptr(pwd.pw_name) })?,
+                _ => Username::nobody(),
+            },
+        };
+
+        let id = match (ret, result.is_null()) {
+            (0, false) => pwd.pw_uid,
+            _ => u32::MAX,
+        };
+
+        if id == 0 {
+            return Err(Error::InvalidState("refusing to authenticate root user"));
+        }
+
+        let gid = match (ret, result.is_null()) {
+            (0, false) => pwd.pw_gid,
+            _ => u32::MAX,
+        };
+
+        let (home_dir, shell) = if ret != 0 {
+            let error = io::Error::from_raw_os_error(ret);
+            debug!(%error, %name, "failed to get user information");
+            (Self::FAKE_HOME, Self::DEFAULT_SHELL)
+        } else if result.is_null() {
+            debug!(%name, "user not found");
+            (Self::FAKE_HOME, Self::DEFAULT_SHELL)
+        } else {
+            // POSIX does not promise these values will be non-null
+            debug!(%name, "found home dir");
+            (
+                match pwd.pw_dir.cast_const() {
+                    home_dir if !home_dir.is_null() => home_dir,
+                    _ => Self::FAKE_HOME,
+                },
+                match pwd.pw_shell.cast_const() {
+                    shell if !shell.is_null() => shell,
+                    _ => Self::DEFAULT_SHELL,
+                },
+            )
+        };
+
+        // SAFETY: if `ret` is 0 (signifying success) and `result` is non-null, `pwd.pw_dir`
+        // and `pwd.pw_shell` were populated by the `getpw` call and the `pwd` struct and `buf`
+        // are still alive, so the pointers are valid; otherwise, `home_dir` and `shell` are set
+        // to static strings. In either case, both are valid pointers to null-terminated C strings.
+        let home_dir = PathBuf::from(OsStr::from_bytes(
+            unsafe { CStr::from_ptr(home_dir) }.to_bytes(),
+        ));
+
+        // An empty `pw_shell` means the system default shell.
+        // SAFETY: `shell` is a valid pointer to a null-terminated C string,
+        // per the same reasoning as for `home_dir` above.
+        let shell = match unsafe { CStr::from_ptr(shell) }.to_bytes() {
+            b"" => PathBuf::from(OsStr::from_bytes(
+                // SAFETY: `DEFAULT_SHELL` points to a static null-terminated C string literal.
+                unsafe { CStr::from_ptr(Self::DEFAULT_SHELL) }.to_bytes(),
+            )),
+            bytes => PathBuf::from(OsStr::from_bytes(bytes)),
+        };
+
+        Ok(User {
+            name,
+            id,
+            gid,
+            home_dir,
+            shell,
+        })
+    }
+
+    const FAKE_HOME: *const c_char = c"/var/empty".as_ptr().cast::<c_char>();
+    const DEFAULT_SHELL: *const c_char = c"/bin/sh".as_ptr().cast::<c_char>();
+}
+
+fn check_permissions(file: &File, uid: u32, level: &str) -> ControlFlow<()> {
+    let meta = match file.metadata() {
+        Ok(meta) => meta,
+        Err(error) => {
+            warn!(%error, level, "failed to get metadata");
+            return ControlFlow::Break(());
+        }
+    };
+
+    match meta.mode() & 0o022 == 0 && (meta.uid() == 0 || meta.uid() == uid) {
+        true => ControlFlow::Continue(()),
+        false => ControlFlow::Break(()),
+    }
 }
 
 /// Element type of the group list passed to `getgrouplist()`, which differs by platform
